refactor(dflash): clean the code
